Add services, repositories, schemas

+++ backend/app/models/product.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, ForeignKey
+from sqlalchemy import Column, Float, Integer, String, ForeignKey
 from sqlalchemy.orm import relationship
 from datetime import datetime
 from app.database import Base
@@ -8,7 +8,7 @@ class Product(Base):
     id = Column(Integer, primary_key=True, index=True)
     name = Column(String, nullable=False)
     description = Column(String)
-    price = Column(Integer, nullable=False)
+    price = Column(Float, nullable=False)
     category_id = Column(Integer, ForeignKey("categories.id"),nullable=False)
     category = relationship("Category", back_populates="products")
     image_url = Column(String)
